Update polyfill and remove token from config
